Make many dependencies optional using Requires.jl
