+++ src/array.jl
@@ -630,16 +630,6 @@ copy!(dest::CatArrOrSub{<:Any, 1}, src::AbstractArray{<:Any, 1}) =
 copy!(dest::CatArrOrSub{T, 1}, src::AbstractArray{T, 1}) where {T} =
     copyto!(dest, 1, src, 1, length(src))
 
-copyto!(dest::CatArrOrSub{<:Any, 1}, src::SentinelArrays.ChainedVector) =
-    copyto!(dest, 1, src, 1, length(src))
-copyto!(dest::CatArrOrSub{<:Any, 1}, dstart::Union{Signed, Unsigned},
-        src::SentinelArrays.ChainedVector, sstart::Union{Signed, Unsigned},
-        n::Union{Signed, Unsigned}) =
-    invoke(copyto!, Tuple{AbstractArray, Union{Signed, Unsigned},
-                          SentinelArrays.ChainedVector,
-                          Union{Signed, Unsigned}, Union{Signed, Unsigned}},
-           dest, dstart, src, sstart, n)
-
 similar(A::CategoricalArray{S, M, R}, ::Type{T},
         dims::NTuple{N, Int}) where {T, N, S, M, R} =
     Array{T, N}(undef, dims)
@@ -1108,33 +1098,6 @@ Base.repeat(a::CatArrOrSub{T, N},
 Base.repeat(a::CatArrOrSub{T, N};
             inner = nothing, outer = nothing) where {T, N} =
     CategoricalArray{T, N}(repeat(refs(a), inner=inner, outer=outer), copy(pool(a)))
-
-# JSON3 writing/reading
-StructTypes.StructType(::Type{<:CategoricalVector}) = StructTypes.ArrayType()
-
-StructTypes.construct(::Type{<:CategoricalArray}, A::AbstractVector) =
-    constructgeneral(A)
-StructTypes.construct(::Type{<:CategoricalArray}, A::Vector) =
-    constructgeneral(A)
-
-function constructgeneral(A)
-    if eltype(A) === Any
-        # unlike `replace`, broadcast narrows the type, which allows us to return small
-        # union eltypes (e.g. Union{String,Missing})
-        categorical(ifelse.(A .=== nothing, missing, A))
-    elseif eltype(A) >: Nothing
-        categorical(replace(A, nothing=>missing))
-    else
-        categorical(A)
-    end
-end
-
-StructTypes.construct(::Type{<:CategoricalArray{Union{Missing, T}}},
-                      A::AbstractVector) where {T} =
-    CategoricalArray{Union{Missing, T}}(replace(A, nothing=>missing))
-StructTypes.construct(::Type{<:CategoricalArray{Union{Missing, T}}},
-                      A::Vector) where {T} =
-    CategoricalArray{Union{Missing, T}}(replace(A, nothing=>missing))
 
 # DataAPI refarray/refvalue/refpool support
 struct CategoricalRefPool{T, P} <: AbstractVector{T}
